requirements.txt
numpy
spidev
lgpio
Pillow
